pyproject.toml
[project]
name = "auto-coin-trader-v2"
version = "0.1.0"
description = ""
authors = [{ name = "jungwoo.mac", email = "qwlake@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "aiohttp (>=3.12.0,<4.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-binance (>=1.0.29,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ujson (>=5.10.0,<6.0.0)",
    "streamlit (>=1.28.0,<2.0.0)",
    "plotly (>=5.17.0,<6.0.0)",
    "pandas (>=2.1.0,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "watchdog>=6.0.0",
    "pydantic-settings>=2.10.1",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
